requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
llama-index == 0.12.40
llama-index-vector-stores-qdrant
qdrant-client==1.14.1
llama-cpp-python==0.2.24
--extra-index-url https://download.pytorch.org/whl/cu118
torchvision==0.22.1
torchaudio==2.7.1
transformers==4.39.3
torch
sentence-transformers==2.7.0
huggingface-hub==0.23.1
llama-index-embeddings-huggingface
fastembed
python-multipart
llama-index-llms-groq
llama-index-readers-file